Crate compose_syntax

Modules

ast

test_utils

Macros

assert_ast

assert_parse_tree

Macro to assert the structure and content of the parsed syntax tree from source code.

assert_tokens

Asserts that the token stream produced by the lexer exactly matches a sequence of expected tokens.

Structs

FileId

Identifier for a file,

Fix

FixBuilder

Label

Lexer

PatchEngine

Source

Span

Defines a range in a source file.

SyntaxError

A syntactical error.

SyntaxNode

Enums

FixDisplay

LabelType

Patch

SyntaxErrorSeverity

SyntaxKind

Functions

parse

Parses one or more statements from the given input text, producing a concrete syntax tree (CST).