compose_syntax

Module test_utils

Macros

assert_ast

assert_parse_tree

Macro to assert the structure and content of the parsed syntax tree from source code.

assert_tokens

Asserts that the token stream produced by the lexer exactly matches a sequence of expected tokens.

Structs

AstTester

NodesTester

Traits

LexerAssert

This is an extension trait for the following impl:

Functions

assert_parse

assert_parse_with_errors

assert_parse_with_warnings

test_file_id

Creates a non-interned file id for testing.

test_parse